Restructuring the internal code to allow for more use cases
